pot/requirements.txt
dnspython
python-dotenv
requests
